Transactions working in SOAP controller
