+++ backend/pom.xml
@@ -17,6 +17,12 @@
 
     <dependencies>
 
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+            <version>1.8.9</version>
+        </dependency>
+
         <!-- Spring framework libraries -->
         <dependency>
             <groupId>org.springframework</groupId>
