+++ backend/src/main/java/pb138/dal/repository/CategoryRepositoryImpl.java
@@ -1,5 +1,7 @@
 package pb138.dal.repository;
 
+import org.springframework.context.annotation.Scope;
+import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import pb138.dal.entities.Category;
@@ -14,11 +16,14 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
+import javax.transaction.Transactional;
 import java.util.LinkedList;
 import java.util.List;
 
 @Component
 @Repository
+@Scope(proxyMode = ScopedProxyMode.INTERFACES)
+@Transactional
 public class CategoryRepositoryImpl implements CategoryRepository {
 
     @PersistenceContext
