+++ web/pom.xml
@@ -16,6 +16,17 @@
     <name>Inventory Management - Web</name>
 
     <dependencies>
+
+        <!--
+        http://stackoverflow.com/questions/37416941/error-creating-bean-in-spring-boot
+        Hibernate transactions for Spring Boot
+        -->
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+            <version>1.8.9</version>
+        </dependency>
+
         <dependency>
             <groupId>org.apache.cxf</groupId>
             <artifactId>cxf-spring-boot-starter-jaxws</artifactId>
